change titles (rename class to match the new requirement)
